Move coma model input output utilities
